requirements.txt
ultralytics<8.2.0,>=8.1.0
opencv-python<4.10.0,>=4.9.0
Pillow<10.3.0,>=10.2.0
numpy<1.25.0,>=1.24.0
torch<2.3.0,>=2.2.0
torchvision<0.18.0,>=0.17.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
matplotlib>=3.7.0,<3.8.0
scikit-learn>=1.3.0,<1.4.0
tqdm>=4.66.0
